Allow demo server to server SharedArrayBuffer wasm
The new implementation for SharedArrayBuffer requires
to be in a secure context. These headers must be present
but it also requires you to either sever it through localhost
or serve it through HTTPS. To test if you are in a secure
context just check window.crossOriginIsolated in javascript.

I think the demo server should be in a gist, or its own
repo, since it's getting more and more useful for
testing purposes. Just like my web services simulator.